DLC: Fix inverted logic for outgoing calls.
The controller uses a "disable outgoing calls" for its storage,
however server passes a "enable outgoing calls" as part of the
configuration.
Currently, the value passed by the server during provisioning
is saved as is by the controller, resulting in an inverted
logic (enabling outgoing calls on the server results in calls
being disabled on the device and vice versa).

Bug: 316418162
Test: atest DeviceLockControllerRoboTests

Change-Id: I77d024d02d030806cc74bd07eae49b474a1835e1