[DLC vNext] Move dependency of TeX to lib
In the previous CLs, I added the TeX dependency to the static_lib
section of the build rules. However this may have caused a duplicate
Java package issue which caused the test in the related bug to fail.
This CL is to fix this issue by moving the dependency from static_lib to
lib, which no longer includes the class into the generated build
targets, thus should solve the issue.

Test: atest CtsStrictJavaPackagesTestCases:android.compat.sjp.cts.StrictJavaPackagesTest#testApkInApex_nonClasspathClasses -- --abi x86_64
Test: atest DeviceLockControllerRoboTests
Test: atest DeviceLockControllerAndroidTests
Bug: b/316909537
Change-Id: I5f7e7311b097362794dc6aa5bd015f4ec6f40673