Refactor device state check up to main method
Currently, we do not check device state until after we verify that the
provision state is in PROVISION_SUCCEEDED. This can be an issue if the
device state is in the CLEARED state and provisioning should not occur.

This CL is a refactor to move the enforcement method logic around so
that both device state and provision state are read before determining
how to enforce the combined states.

There should be no functional changes in this CL, just a refactor. I
will follow-up with a CL with the actual fix.

Bug: 317418278
Test: atest DeviceLockControllerRoboTests
Change-Id: Ic11e6fe5d540c59de871553b3ca560410b2d69cb