Send notification on receiving provision state
We currently always send the notification *after* the alarm triggers to
go to the next provision state (e.g. 1 day after entering the state).

This does not really work because the notification should be shown
while the state is active, not when we are about to move to the next
state.

Instead, send the notification after entering the provision state at the
same time as we schedule the alarm for the next state.

This removes the method to run the alarm receiver immediately as it is
no longer necessary. (Seemed to be a workaround to try and trigger the
notification sending immediately on entering the state but had the
unwanted side-effect of also progressing to the next state immediately).

Bug: 322047488
Test: atest DeviceLockControllerRoboTests
Change-Id: I30c29bfd2a11bb6bef670dbe2019448f9b51a228