Only disable DLC for user after clean up
Do not automatically disable the DLC package when the device is
finalized. Instead, we should wait for the DLC package to finish its
clean-up (e.g. roles, permissions, etc.) and then tell us to disable
it.

The only exception to this is the system user which needs to have its
DLC package enabled as long as other users have not finished their
finalization tasks. This is b/c the system user DLC package hosts
services that non-system users need to use to clean up (e.g. global
parameters). This can be safely disabled after all users are disabled as
any new users will start with the DLC package disabled.

To summarize:

If device is not finalized
=> Always enable DLC on user switch / unlock

If device is finalized
  If not system user
    => Only disable on user switch / unlock after clean up finishes
    => Always disable immediately for new user
  If system user
    => Only disable on user switch / unlock after clean up finishes
       AND all other user DLCs are disabled

Bug: 328110278
Test: atest DeviceLockUnitTests
Change-Id: I8b1ee006504d9b7afaa57b737337cd36fd4ea369