Re-enable ignored tests
Tests were failing because we could not wait for the future to complete
before making our test assertions because the executor used was opaque
to the test and not sequential.

This CL does a small refactor so that we can inject a single-threaded
executor that the test can wait on to become idle before continuing.

Bug: 327652632
Test: atest DeviceLockControllerRoboTests
Change-Id: I1b0c11e1209623a49c732e7504447d0d5451f823