Fix child being focused separately from section
"Backing up and restoring data" child text was focusable separately from
the containing parent because it had a URL. This caused it to be read
separately from the other examples in the section. By making it not
focused but keeping the parent focused, it gets read with the rest of
the examples (the link is still preserved in the accessibility info of
the full section).

Also add unit tests for DevicePolicyGroupViewHolder

Bug: 328119031
Test: atest DeviceLockControllerRoboTests
Change-Id: I7b9e10d4fe80478d1a83c99ed3fb5ee2fb0a0691