Fallback to non-VPN if default network fails
This CL modifies the check-in client to attempt to use a non-VPN network
if the call fails on the default network. This only occurs if there is a
validated non-VPN network.

We also add a 10 second gRPC timeout to all calls so that we do not hang
on a network call.

Bug: 332905713
Test: repro in bug
Change-Id: I3cc0bd1dcee45abef283d71cfd9bf3679f2ac245