Add non-VPN to networked work constraints
The workers were starting when the VPN was connected even when it had
no underlying transport which caused it to fail since there was no
actual connectivity.

We can simply add non-VPN to the constraints since we know that a VPN
with connectivity should have an underlying network with connectivity.

Note that the actual networked call itself will start with the default
network which may be the VPN network.

Bug: 337058538
Test: atest DeviceLockControllerRoboTests
Test: manual
Change-Id: Idafc6c88157cfbfd7a20bf0bc2ade92d95dfef30