Fix exception logging line not being run
The log line logging the exception in the case of a policy enforcement
failure never ran because the handler logic would cancel the work
running the logging code.

This moves the log up a level so that the exception is logged *before*
the handler code runs and cancels the work.

Bug: 341158221
Test: builds
Change-Id: I258f0a90df70972e1698aa96d5b89163ab04892e