Start lock task mode if launching lock intent
If the DLC app dies while lock task mode is active, pressing the home
button will exit lock task mode. This is because we have
"initialization" logic when getting the lock task intent that will
enforce the current policies and one of the policies we enforce will
*unlock* the device in preparation for locking. However, because we are
just getting the intent, we don't do the usual full locking logic which
just ends up with DLC unlocking only.

This CL addresses this by adding the logic to lock the device after as
well.

Bug: 340313786
Flag: EXEMPT bugfix
Test: atest DeviceLockControllerRoboTests
Change-Id: I76177be3fb6f1f76cdf57041faac2db189a483bd