Ensure clear restrictions is terminal.
If a kiosk app sends a clear() command followed immediately by an unlock() command, sometimes (about 1 in 5 times), the unlocked state will be enforced instead of the clear state. This is due to a race condition in DeviceStateControllerImpl.
This CL ensures that a clear() command is always terminal and not interfered with by this race condition by introducing a clearingInProgress
variable that is set the moment a clear() command is received.

Bug: 379925834
Test: Verified manually that the issue does not recur
Flag: EXEMPT bugfix
Change-Id: Ica0bd23cf1bb1b92176e958a7456266ada0131cc